Descargar

Descarga datos disponibles del portal de Microdatos del INEI.

Usage

descargar.inei(
  encuesta = "ENAHO",
  modulo,
  periodos,
  dirdescarga = getwd(),
  tipo = "anual",
  ensilencio = FALSE
)

Arguments

encuesta

el nombre de una encuesta del INEI. Las opciones dispobibles son "ENAHO" y "ENAHOpanel". No importa si está escrito en mayúscula o minúscula.

modulo

el código del módulo deseado.

periodos

los años deseados.

dirdescarga

el directorio de descarga.

tipo

el tipo de período deseado. Las opciones son "anual", "t1" (primer trimestre), "t2" (segundo trimestre), "t3" (tercer trimestre), y "t4" (cuarto trimestre). Las opciones pueden combinarse.

ensilencio

un valor lógico que indica si el progreso de la función debe ser omitido.

Value

guarda los archivos descargados en el disco.

Examples

descargar.inei(modulo = 37, periodos = 2011, dirdescarga = tempdir(), tipo = "t1")
#> Descargando 1 archivo.