Leer bases de la ENAHO

Lee bases de la ENAHO descargadas con descargar.inei. Si más de una base es leída, se carga una lista, cuyos elementos son cada una de las bases. En caso que el INEI provea más de una base por módulo y período se leerá la base con mayor peso.

Usage

leer.inei(
  encuesta = "ENAHO",
  modulo,
  periodos,
  directorio = getwd(),
  tipo = "anual",
  ensilencio = FALSE,
  combinar = FALSE,
  solocomunes = FALSE,
  columnas = NULL
)

Arguments

encuesta

el nombre de una encuesta del INEI. Las opciones dispobibles son "ENAHO" y "ENAHOpanel". No importa si está escrito en mayúscula o minúscula.

modulo

el código del módulo deseado.

periodos

los años deseados.

directorio

el directorio de donde están las carpetas generadas por descargar.inei.

tipo

el tipo de período deseado. Las opciones son "anual", "t1" (primer trimestre), "t2" (segundo trimestre), "t3" (tercer trimestre), y "t4" (cuarto trimestre). Las opciones pueden combinarse.

ensilencio

un valor lógico que indica si el progreso de la función debe ser omitido.

combinar

un valor lógico que indica si las bases deben ser combinadas en un único data frame.

solocomunes

un valor lógico que indica si sólo deben combinarse las columnas comunes.

Value

un data frame o una lista.

Examples

directorio = system.file("extdata", package = "enaho")
leer.inei(modulo = 37, periodos = 2011, directorio = directorio, tipo = "t1")
#> Leyendo 1 archivo.
#> Leyendo archivo 1 de 1.
#> # A tibble: 4,625 × 24
#>    AÑO   MES   CONGLOME VIVIENDA HOGAR UBIGEO DOMINIO ESTRATO CODINFOR  P702
#>    <chr> <chr> <chr>    <chr>    <chr> <chr>    <dbl>   <dbl> <chr>    <dbl>
#>  1 2011  01    0007     049      11    010101       4       4 01           3
#>  2 2011  01    0007     075      11    010101       4       4 02           5
#>  3 2011  01    0027     047      11    010504       4       7 02           2
#>  4 2011  01    0027     047      11    010504       4       7 02           3
#>  5 2011  01    0027     047      11    010504       4       7 02           5
#>  6 2011  01    0027     047      11    010504       4       7 02           5
#>  7 2011  01    0027     108      11    010504       4       7 02           3
#>  8 2011  01    0027     108      11    010504       4       7 02           4
#>  9 2011  01    0027     109      11    010504       4       7 01           4
#> 10 2011  01    0027     110      11    010504       4       7 02           4
#> # ℹ 4,615 more rows
#> # ℹ 14 more variables: P703 <dbl+lbl>, P704 <dbl+lbl>, P705 <dbl+lbl>,
#> #   P7061 <dbl+lbl>, P7062 <dbl+lbl>, P7063 <dbl+lbl>, P7064 <dbl+lbl>,
#> #   P7065 <dbl+lbl>, P706A1 <dbl+lbl>, P706A2 <dbl+lbl>, P706A3 <dbl+lbl>,
#> #   P706A4 <dbl+lbl>, TICUEST01 <dbl+lbl>, FACTRIM <dbl>