Media

Estima medias de variables en bases leídas con leer.inei o leer.inei.web.

Usage

media.enaho(
  x,
  base,
  estratos = NULL,
  combinarestratos = FALSE,
  dividirperiodos = FALSE,
  ruralidad = 6:8,
  pesos = "FACTOR07",
  var = c("insesgado"),
  decimales = NULL
)

Arguments

x

una cadena de caracteres con los nombres de las variables a estimar.

base

una base leída con leer.inei o leer.inei.web en forma de tibble o en forma de lista.

estratos

una cadena de caracteres con los nombres de las variables que deben ser consideradas como estratos. Además se puede usar "area" y "region", con las que se calcularán los resultados para los estratos urbano y rural, y para cada región.

combinarestratos

un valor lógico que indica si debe calcularse también la combinación entre estratos. Por ejemplo, si los estatos son "area" y "region", se estimará también para cada área de cada región.

dividirperiodos

un valor lógico que indica si los resultados deben ser separados por período.

ruralidad

un vector numérico que indica que valores de la variable "ESTRATO" deben ser considerados rurales. Usualmente del 6 al 8.

pesos

una cadena de caracteres con el nombre de la(s) variable(s) que representan los pesos. Si es NULL las proporciones se estiman sin pesarlas.

var

una cadena de texto que indica el método a usar para la varianza: "insesgado" calcula la estimación insesgada (n-1); "MV" calcula la estimación de máxima verosimilitud.

decimales

un valor numérico que indica cuántos decimales se necesitan. Si es NULL no se redondea.

Value

un data frame o una lista.

Examples

directorio = system.file("extdata", package = "enaho")
leido <- leer.inei(modulo = 37, periodos = 2011, directorio = directorio, tipo = c("t1","t2"))
#> Leyendo 2 archivos.
#> Leyendo archivo 1 de 2.
#> Leyendo archivo 2 de 2.
media.enaho(x = c("P7061"),base = leido,pesos = "FACTRIM")
#>   Período  Estrato   Nombre Variable     Media        DE  Varianza
#> 1 2011_t1 Nacional Nacional    P7061 0.4905288 0.4999103 0.2499104
#> 2 2011_t2 Nacional Nacional    P7061 0.4647675 0.4987572 0.2487587