Proporción

Estima proporciones de variables en bases leídas con leer.inei o leer.inei.web.

Usage

proporcion.enaho(
  x,
  base,
  estratos = NULL,
  combinarestratos = FALSE,
  dividirperiodos = FALSE,
  ruralidad = 6:8,
  pesos = "FACTOR07",
  forzaretiquetas = FALSE,
  porcentaje = FALSE,
  decimales = NULL,
  formatolargo = FALSE
)

Arguments

x

una cadena de caracteres con los nombres de las variables a estimar.

base

una base leída con leer.inei o leer.inei.web en forma de tibble o en forma de lista.

estratos

una cadena de caracteres con los nombres de las variables que deben ser consideradas como estratos. Además se puede usar "area" y "region", con las que se calcularán los resultados para los estratos urbano y rural, y para cada región.

combinarestratos

un valor lógico que indica si debe calcularse también la combinación entre estratos. Por ejemplo, si los estatos son "area" y "region", se estimará también para cada área de cada región.

dividirperiodos

un valor lógico que indica si los resultados deben ser separados por período.

ruralidad

un vector numérico que indica que valores de la variable "ESTRATO" deben ser considerados rurales. Usualmente del 6 al 8.

pesos

una cadena de caracteres con el nombre de la variable que representa los pesos. Si es NULL las proporciones se estiman sin pesarlas.

forzaretiquetas

un valor lógico que indica si las etiquetas del período con más etiquetas deben ser forzadas para todos los períodos. No es poco común que las etiquetas de una variable cambien con los años, si esto ocurre, esta función las interpretará como distintas. Por ejemplo "INCIAL" y "EDUCACION INICIAL". Sin embargo, usando este argumento se puede forzar a que se interpreten los mismos valores todos los años. Usar con discreción, y, si es posible, siempre después de una exploración de datos usando forzaretiquetas = FALSE.

porcentaje

un valor lógico que indica si los resultados deben presentarse como porcentaje en vez de proporción.

decimales

un valor numérico que indica cuántos decimales se necesitan. Si es NULL no se redondea.

formatolargo

un valor lógico que indica si los resultados debes estar en formato largo. Si se incluye más de una variable, se darán los resultados en formato largo.

Value

un data frame o una lista.

Examples

directorio = system.file("extdata", package = "enaho")
leido <- leer.inei(modulo = 37, periodos = 2011, directorio = directorio, tipo = c("t1","t2"))
#> Leyendo 2 archivos.
#> Leyendo archivo 1 de 2.
#> Leyendo archivo 2 de 2.
proporcion.enaho(x = c("P7061"),base = leido,pesos = "FACTRIM")
#>   Período  Estrato   Nombre Variable      Pase Sí, pagó en dinero
#> 1 2011_t1 Nacional Nacional    P7061 0.5094712          0.4905288
#> 2 2011_t2 Nacional Nacional    P7061 0.5352325          0.4647675