Identificar variables

Identifica las variables disponibles por periodo y tipo de bases de la ENAHO descargadas con descargar.inei.

Usage

variablesxperiodo(
  encuesta = "ENAHO",
  modulo,
  periodos,
  directorio = getwd(),
  tipo = "anual",
  ultimaetiqueta = TRUE,
  mostraretiqueta = FALSE
)

Arguments

encuesta

el nombre de una encuesta del INEI. Las opciones dispobibles son "ENAHO" y "ENAHOpanel". No importa si está escrito en mayúscula o minúscula.

modulo

el código del módulo deseado.

periodos

los años deseados.

directorio

el directorio de donde están las carpetas generadas por descargar.inei.

tipo

el tipo de período deseado. Las opciones son "anual", "t1" (primer trimestre), "t2" (segundo trimestre), "t3" (tercer trimestre), y "t4" (cuarto trimestre). Las opciones pueden combinarse.

ultimaetiqueta

un valor lógico que indica si debe usarse la última etiqueta. Dado que la etiqueta de las variables puede cambiar con los años, se puede elegir qué etiqueta usar, la del período más reciente que tiene esa variable ultimaetiqueta = TRUE o del primer período que tiene esa variable ultimaetiqueta = FALSE.

mostraretiqueta

un valor lógico que indica si deben mostrarse las etiquetas.

Value

un data frame donde 1 significa que la variable está disponible y 0 que no está disponible.

Examples

directorio = system.file("extdata", package = "enaho")
variablesxperiodo(modulo = 37, periodos = 2011, directorio = directorio, tipo = c("t1","t2"))
#>       Nombre 2011_t1 2011_t2
#> 1        AÑO       1       1
#> 2        MES       1       1
#> 3   CONGLOME       1       1
#> 4   VIVIENDA       1       1
#> 5      HOGAR       1       1
#> 6     UBIGEO       1       1
#> 7    DOMINIO       1       1
#> 8    ESTRATO       1       1
#> 9   CODINFOR       1       1
#> 10      P702       1       1
#> 11      P703       1       1
#> 12      P704       1       1
#> 13      P705       1       1
#> 14     P7061       1       1
#> 15     P7062       1       1
#> 16     P7063       1       1
#> 17     P7064       1       1
#> 18     P7065       1       1
#> 19    P706A1       1       1
#> 20    P706A2       1       1
#> 21    P706A3       1       1
#> 22    P706A4       1       1
#> 23 TICUEST01       1       1
#> 24   FACTRIM       1       1